move user_manager function to InternalServer as suggested by @oroulet
